[Internal] Disable all logs when invoke a flow with Flow.load (#2865)
# Description

Currently we would print logs when directly using the following code:
```python
f = Flow.load(some_flow)
print(f(input_str="some_str"))
```

![image](https://github.com/microsoft/promptflow/assets/3871260/64eaad3b-40ed-4771-8015-8dff7dfc1d10)


After this change, no log is printed.

![image](https://github.com/microsoft/promptflow/assets/3871260/68f69a94-6072-4998-99a0-782ef79bc5ce)

This pull request primarily focuses on improving logging in the
`promptflow-core` package. The changes include adding a `contextlib`
import and implementing a method to disable all loggers in the
`LoggerFactory` class. Additionally, the `invoke` method in the `Flow`
class was modified to disable all loggers while it is being executed.

Improvements to logging:

*
[`src/promptflow-core/promptflow/_utils/logger_utils.py`](diffhunk://#diff-6e8d63d4321bccef88f94d46fa9a0808b373b2024791fbdc732ec11b41938eb9L8-R8):
Added an import for `contextlib` to enable the creation of a context
manager.
*
[`src/promptflow-core/promptflow/_utils/logger_utils.py`](diffhunk://#diff-6e8d63d4321bccef88f94d46fa9a0808b373b2024791fbdc732ec11b41938eb9R375-R383):
Added a list of `_existing_loggers` and a `_disabled` attribute to the
`LoggerFactory` class. Modified the `get_logger` method to add new
loggers to the `_existing_loggers` list and set their `disabled`
attribute based on the `_disabled` attribute of the `LoggerFactory`.
*
[`src/promptflow-core/promptflow/_utils/logger_utils.py`](diffhunk://#diff-6e8d63d4321bccef88f94d46fa9a0808b373b2024791fbdc732ec11b41938eb9R414-R431):
Implemented the `disable_all_loggers` method in the `LoggerFactory`
class as a context manager that disables all existing loggers when
entered and re-enables them when exited.

Changes to flow invocation:

*
[`src/promptflow-core/promptflow/core/_flow.py`](diffhunk://#diff-e40a64a476472cf8da046f0140db9ba9d8e2b47269333e0d7f9fd3d01c976c1eR131-R136):
Modified the `invoke` method in the `Flow` class to disable all loggers
while it is being executed. The actual invocation was moved to a new
`_invoke` method.

# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.

Co-authored-by: Heyi <heta@microsoft.com>