[Doc]Update flex flow doc  (#3075)
# Description

Please add an informative description that covers that changes made by
the pull request and link all relevant issues.
This pull request includes changes to the
`docs/how-to-guides/develop-a-flex-flow` directory, specifically the
`class-based-flow.md` and `function-based-flow.md` files. The changes
primarily involve the addition of new configurations and parameters to
the examples provided in these files.

*
[`docs/how-to-guides/develop-a-flex-flow/class-based-flow.md`](diffhunk://#diff-0aa42d1976c0f4db59cb05a81413c6dafc5f6c6bb6f9c82961299aad956fd2e6R97-R108):
Added `AzureOpenAIModelConfiguration` to the example and updated the
`pf.run` method to include the `model_config` and `flow_config`
parameters. This provides a more detailed example of how to run a flow
class directly with specific configurations.
*
[`docs/how-to-guides/develop-a-flex-flow/function-based-flow.md`](diffhunk://#diff-af24f978caaac273a3a303a2271b7bd374c55c87bf6c1bca283b796df3efd31bR96):
Added `inputs` to the `sample` section of the flow YAML example. This
shows how to specify inputs in a flow YAML file.

# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.