[Internal][tracing] Refactor span enrich logic to make it more clear. (#3056)
# Description

Refactor span enrich logic to make it more clear.

This pull request introduces a new structure for enriching spans in the
tracing module of the `promptflow-tracing` package. The changes
primarily involve the creation of a `SpanEnricher` class and a
`SpanEnricherManager` singleton class in the
`src/promptflow-tracing/promptflow/tracing/_span_enricher.py` file.
These classes are used to enrich spans with inputs and outputs of traced
functions. Two specific enrichers, `LLMSpanEnricher` and
`EmbeddingSpanEnricher`, are created and registered with the manager.
The `TraceType` enum is also updated to include a new `RETRIEVAL` type.

Key changes include:

*
[`src/promptflow-tracing/promptflow/tracing/_span_enricher.py`](diffhunk://#diff-695282eb7072d912e1bbe30979ce4b19b7c4ffe45a76fd64d48d1c392d4f5c25R1-R49):
Introduced `SpanEnricher` and `SpanEnricherManager` classes. The manager
is a singleton that maintains a dictionary mapping trace types to their
respective span enrichers. The base `SpanEnricher` class provides a
method to enrich a span with the output of a traced function.

*
[`src/promptflow-tracing/promptflow/tracing/_trace.py`](diffhunk://#diff-580941184737186a94e3e9c06e467e86c06ce846d30ffa42c1c43b45812ddcdcR23):
Imported the `SpanEnricher` and `SpanEnricherManager` classes. The
`enrich_span_with_trace_type` function now uses the
`SpanEnricherManager` to enrich spans based on their trace type. Two
specific enrichers, `LLMSpanEnricher` and `EmbeddingSpanEnricher`, are
implemented and registered with the `SpanEnricherManager`.
[[1]](diffhunk://#diff-580941184737186a94e3e9c06e467e86c06ce846d30ffa42c1c43b45812ddcdcR23)
[[2]](diffhunk://#diff-580941184737186a94e3e9c06e467e86c06ce846d30ffa42c1c43b45812ddcdcL150-R154)
[[3]](diffhunk://#diff-580941184737186a94e3e9c06e467e86c06ce846d30ffa42c1c43b45812ddcdcR515-R532)

*
[`src/promptflow-tracing/promptflow/tracing/contracts/trace.py`](diffhunk://#diff-064c539f55a396428c9d28af1cbcf4fe39beb69daac7d9891cd6f713897fa704R18):
Added `RETRIEVAL` as a new trace type to the `TraceType` enum.

These changes aim to make the process of enriching spans more flexible
and extensible, allowing for different enrichment logic based on the
trace type.

# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.

Co-authored-by: Heyi <heta@microsoft.com>