feat: pf flow serve support for directory with 1 and only 1 prompty (#3235)
# Description

Support passing a directory without `flow.dag.yaml` nor `flow.flex.yaml`
but with 1 and only 1 prompty in it as `source` in `pf flow serve`.


![image](https://github.com/microsoft/promptflow/assets/37076709/b9e67e44-1955-40dc-a417-a5402a1c5aa7)

App support is already done by support prompty in `init_executable`, so
this PR is mainly focus on changes in cli.

Note that this is not a full support for prompty in `pf flow serve` as
there can usually be multiple prompty in 1 directory. We will need to
figure out this in next release.

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.