[Bugfix][Executor] Persist multimedia data after getting a LineResult from output queue in line process pool for execution server (#3231)
# Description

### Background
When testing the resume feature in the multi-container, it was
discovered that the resume run fails to correctly load images in the
api_calls of the previous run's flow run info. The underlying issue is
that the api_calls in the flow run are assigned the api_calls of the
child node runs. However, at this point, the images have been persisted
as paths under node artifacts. Therefore, the images cannot be found
under flow artifacts, leading to an error.

### Fix solution

- Use persist_multimedia_after_execution to decide persist images after
execution or during execution, and set it to True for execution server.
- Remove ServiceQueueRunStorage and do not persist images during real
execution.
- Use service storage to persist images after getting a LineResult from
the output queue.

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.