Fix the recordings to allow promptflow runs. (#3477)
# Description

The promptflow creates a lot of temporary files on-the -ground, which
are being transferred to the cloud. In this PR I have added the code to
sanitize the run names and added other fixes to allow run recordings.
1. The run name and experiment name are now sanitized as well as other
ID-s.
2. Added the sanitized tracking URI in the recordings
3. PF run first checks if the run with the same name exists and if it
does it gets 404. If we continue run, we are getting two different
responses to the same run name request: first 404 and then 200 (when the
run was already created). These requests are the same and if failed with
404 request will be recorded, the second i.e. continuing run will fail.
I have added the code to overwrite only failed runs (code >= 400)
4. The resulting cassettes were re recorded.

See work item 3305909

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [x] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [x] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.