Bugfix/non azure open ai prompty (#3621)
# Description

Fix evaluators to accept model configs of type
`OpenAIModelConfiguration`. Highlights of this change include:
- A bunch of typehint updates.
- The actual bugfix: Making a common azure-openai-config check only
apply to configs of type `AzureOpenAIModelConfiguration` instead of all
inputted configs.
- Secondary bugfix: Make `validate_interface` method properly handle
both singleton and list outputs of the `resolve_annotation` method.
- Added a test to the QA eval (since this implicitly calls all other
evaluators). This test is currently disabled since we don't have a
team-wide openAI API key to use for general testing yet, but I had it
passing locally when using a personal API key.
- Added an OpenAIConfiguration fixture that's also currently un-usable
for the same reason.